requirements.txt
requests
Pillow
pytesseract
opencv-python
numpy
mss
